+++ lib/ts/recipe/session/utils.ts
@@ -18,7 +18,6 @@ import {
     TypeInput,
     TypeNormalisedInput,
     NormalisedErrorHandlers,
-    InputSchema,
     ClaimValidationError,
 } from "./types";
 import {
@@ -35,7 +34,7 @@ import { REFRESH_API_PATH } from "./constants";
 import NormalisedURLPath from "../../normalisedURLPath";
 import { NormalisedAppinfo } from "../../types";
 import * as psl from "psl";
-import { isAnIpAddress, validateTheStructureOfUserInput } from "../../utils";
+import { isAnIpAddress } from "../../utils";
 import { RecipeInterface, APIInterface } from "./types";
 import { BaseRequest, BaseResponse } from "../../framework";
 import { sendNon200Response } from "../../utils";
@@ -145,7 +144,6 @@ export function validateAndNormaliseUserInput(
     appInfo: NormalisedAppinfo,
     config?: TypeInput
 ): TypeNormalisedInput {
-    validateTheStructureOfUserInput(config, InputSchema, "session recipe");
     let cookieDomain =
         config === undefined || config.cookieDomain === undefined
             ? undefined